+++ src/policybuilder.rs
@@ -172,6 +172,9 @@ impl<T: PolicyWriter> PolicyBuilder<T> {
                 );
             }
 
+            // TODO: This is writing the old format of just DOMAIN.KEY.  We
+            //       need to now keep the mapping in the policy.
+            // TODO: Also the "identity attrs" is now using the service names, not the zpl names.
             if let Some((l7p, port)) = svc.get_l7protocol_and_port() {
                 self.policy_writer.write_trusted_service(
                     &canonical_svc_id,